[REF] base_automation: lint cleanup, fix stale machine_doc reference
Loop A audit (t24068). Read machine_doc_v1/ first per repo convention -
conventions.md/models.md already document several known gotchas
(recursion guard, safe_eval scope, cron-interval-only-decreases) that
were verified still accurate, so not re-litigated here.

Security review: test_webhook_security.py already covers HMAC valid/
invalid, IP allowlist, payload size limit, and none-auth-open-by-
default - webhook hardening has clearly already had a dedicated pass;
no new gap found on read-through of controllers/main.py and the
webhook execution path in base_automation.py.

Fixed one incidental find: models.md's DAG-resolution section still
contrasted automation.runtime.line.action_mark_done() against
"ir.actions.server.action_mark_done() which mutates the global
definition" - that method was removed in Phase 1 (see conventions.md's
"What NOT to Add" section) and no longer exists; the doc reference was
stale. Corrected.

Code changes are cosmetic only (import ordering, RET503, TRY201
raise-without-name, RUF046 redundant int(round()), SIM105 contextlib.
suppress, unused imports/variables in tests) plus two `# noqa: B023`
on lambdas that ruff flags as "doesn't bind loop variable" but are
verified false positives: both are consumed immediately by .filtered()
within the same loop iteration, never stored for delayed execution.

Note: an initial attempt at the TRY201 fix used a global find-replace
of the literal text "raise e" and silently corrupted six unrelated
`raise exceptions.ValidationError(...)` call sites into
`raisexceptions.ValidationError(...)` (matched as a substring) -
caught immediately by the next lint pass (F821 undefined name) and
fixed before running any tests. Worth remembering: never blind
find-replace short exception-raising snippets in a file that also
raises via `exceptions.X(...)` under a bare `import ... exceptions`.

3 pre-existing test issues confirmed unrelated via git-stash roundtrip
(2 browser-tour flakes in test_mail_composer.py, 1 setUpClass error in
test_webhook_security.py needing ODOO_API_ENCRYPTION_KEY which isn't
set in this shell) - identical on the unmodified baseline.

Task ID: 24068

+++ addons/base_automation/models/automation_runtime.py
@@ -179,7 +179,7 @@ class AutomationRuntime(models.Model):
                 runtime.progress = 0
                 continue
             done = len(runtime.line_ids.filtered(lambda l: l.state == "done"))
-            runtime.progress = int(round((done / total) * 100))
+            runtime.progress = round((done / total) * 100)
 
     @api.depends("line_ids.state")
     def _compute_progress_display(self):
